Add Rob's condensing patch
