+++ src/ipahealthcheck/ds/config.py
@@ -2,8 +2,6 @@
 # Copyright (C) 2020 FreeIPA Contributors see COPYING for license
 #
 
-from ipahealthcheck.core import constants
-from ipahealthcheck.core.plugin import Result, duration
 from ipahealthcheck.ds.plugin import DSPlugin, registry
 from lib389.config import Config
 
@@ -13,13 +11,4 @@ class ConfigCheck(DSPlugin):
     """
     Check the DS config for obvious errors
     """
-    requires = ('dirsrv',)
-
-    @duration
-    def check(self):
-        results = self.doCheck(Config)
-        if len(results) > 0:
-            for result in results:
-                yield result
-        else:
-            yield Result(self, constants.SUCCESS)
+    check_class = Config
