+++ src/ipahealthcheck/ds/disk_space.py
@@ -2,8 +2,6 @@
 # Copyright (C) 2020 FreeIPA Contributors see COPYING for license
 #
 
-from ipahealthcheck.core import constants
-from ipahealthcheck.core.plugin import Result, duration
 from ipahealthcheck.ds.plugin import DSPlugin, registry
 from lib389.monitor import MonitorDiskSpace
 
@@ -13,13 +11,4 @@ class DiskSpaceCheck(DSPlugin):
     """
     Check the all the disks that the DS uses
     """
-    requires = ('dirsrv',)
-
-    @duration
-    def check(self):
-        results = self.doCheck(MonitorDiskSpace)
-        if len(results) > 0:
-            for result in results:
-                yield result
-        else:
-            yield Result(self, constants.SUCCESS)
+    check_class = MonitorDiskSpace
